Get rid of ObservationProvider, move useObservations hook
