+++ src/components/Observations/ObsList.js
@@ -1,12 +1,11 @@
 // @flow
 
-import React, { useContext, useEffect } from "react";
+import React, { useEffect } from "react";
 import type { Node } from "react";
 import { Pressable, Text } from "react-native";
 import { useRoute } from "@react-navigation/native";
 
 import ViewWithFooter from "../SharedComponents/ViewWithFooter";
-import { ObservationContext } from "../../providers/contexts";
 import ObservationViews from "../SharedComponents/ObservationViews/ObservationViews";
 import UserCard from "./UserCard";
 import { useCurrentUser } from "./hooks/useCurrentUser";
@@ -14,10 +13,11 @@ import BottomModal from "../SharedComponents/BottomModal";
 import RoundGreenButton from "../SharedComponents/Buttons/RoundGreenButton";
 import uploadObservation from "../../providers/helpers/uploadObservation";
 import Observation from "../../models/Observation";
+import useObservations from "./hooks/useObservations";
 
 const ObsList = ( ): Node => {
   const { params } = useRoute( );
-  const { observationList, loading, syncObservations, fetchNextObservations, obsToUpload } = useContext( ObservationContext );
+  const { observationList, loading, syncObservations, fetchNextObservations, obsToUpload } = useObservations( );
 
   const id = params && params.userId;
